backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for realtime transcription service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich[toml]>=0.9.2",
    "fastapi[standard]>=0.123.0",
    "pydantic-settings>=2.12.0",
    "pytest-asyncio>=1.3.0",
    "vosk>=0.3.45",
]

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
